requirements.txt
ccxt>=2.0.0
numpy>=1.22.0
pandas>=1.4.0
arch>=5.0.0
python-dotenv>=0.20.0
matplotlib>=3.5.0